Refactored numerical constants
 - distinguish different kind of constants (float vs double), (signed vs unsigned), (long vs long long vs regular)
 - parse Char and w_char
 - fix lexer a bit

+++ include/parser.h
@@ -31,7 +31,7 @@ typedef struct _Token {
     int rawCode;
     const char *text;
     union {
-      long long iv;
+      int64_t iv;
       double dv;
     } value;
     struct _Token *next;
